requirements.txt
groq
streamlit
pillow==11.0.0
openai==1.58.1
requests==2.32.3
python-dotenv==1.0.1
huggingface_hub